+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MediaSegment and Fragment have new StartPos attribute
 - mp4.File now has Mfra pointer
 - MfraBox has new method FindEntry
+- MediaSegment, Fragment, and Trun method CommonSampleDuration
 
 ### Fixed